Created an arrow sprite sheet.
The arrow points in the correct direction depending on the direction of the player.
I think I solved the arrow does not shoot in the direction of the player bug by making sure velocity is set to zero when the arrow hits a wall or enemy.
This introduced some messy code that needs to be cleaned up.